refactor addProject code to helpers/addProjectForm.js
